Split psd-tools code into separate file to fix issue with using types
